missing comments added, unused code removed
