+++ src/frontend/src/services/DataService.ts
@@ -7,20 +7,37 @@ import { CheckResultSet } from '@/models/CheckResultSet';
 
 export class DataService {
 
-  private get baseUrl() {    
+  /**
+   * Return base url for api access.
+   *
+   * @readonly
+   * @private
+   * @memberof DataService
+   */
+  private get baseUrl()   
+  {    
     return ConfigService.ApiUrl;
   }
 
+  /**
+   * Return current api version.
+   *
+   * @readonly
+   * @private
+   * @type {string}
+   * @memberof DataService
+   */
   private get apiVersion():string {
     return '0.1'
   }
 
-  public fixedEncodeURIComponent(str: string) {
-    return encodeURIComponent(str).replace(/[!*]/g, function (c) {
-      return "%" + c.charCodeAt(0).toString(16);
-    });
-  }
-
+  /**
+   * Return general overview data.
+   *
+   * @param {DateTime} [date]
+   * @returns {(Promise<void | InfrastructureOverview>)}
+   * @memberof DataService
+   */
   public async getGeneralOverviewData(date?: DateTime): Promise<void | InfrastructureOverview> {
 
     let suffix = (date === undefined) ? '?api-version=' + this.apiVersion 
@@ -31,14 +48,21 @@ export class DataService {
 
     return axios
       .get(url)
-      .then((response) => {
-         console.log(response.data);
-         return response.data
-      })
+      .then((response) => response.data)
       .then((data) => InfrastructureOverview.GenerateFromData(data))
-      //.catch((error) => console.log(error));
   }
 
+  /**
+   * Return general overview detail.
+   *
+   * @param {number} pageSize
+   * @param {number} pageIndex
+   * @param {DateTime} [date]
+   * @param {string} [filterBy]
+   * @param {string} [sortBy]
+   * @returns {(Promise<void | CheckResultSet>)}
+   * @memberof DataService
+   */
   public async getGeneralOverviewDetail(pageSize: number, pageIndex: number, date?: DateTime, filterBy?: string, sortBy?: string) : Promise<void | CheckResultSet> {
     let suffix = (date === undefined) ? '?api-version=' + this.apiVersion 
                                       : '?date=' + date!.toISODate() + '&api-version=' + this.apiVersion;
@@ -62,6 +86,14 @@ export class DataService {
             .then((data) => <CheckResultSet>data);
   }
 
+  /**
+   * Return general overview search/filtering data.
+   *
+   * @param {DateTime} [date]
+   * @param {string} [filterBy]
+   * @returns {(Promise<void | any>)}
+   * @memberof DataService
+   */
   public async getGeneralOverviewSearch(date?: DateTime, filterBy?: string) : Promise<void | any> {
     let suffix = (date === undefined) ? '?api-version=' + this.apiVersion 
                                       : '?date=' + date!.toISODate() + '&api-version=' + this.apiVersion;
@@ -73,12 +105,18 @@ export class DataService {
     }  
     let url = this.baseUrl + "/audits/overview/search/" + suffix;
     console.log(`[] calling ${url}`);
-
     return axios.get(url)
-            .then((response) => response.data)
-            //.then((data) => <CheckResultSet>data);
+                .then((response) => response.data)
   }
 
+  /**
+   * Return component detail data.
+   *
+   * @param {string} id
+   * @param {DateTime} [date]
+   * @returns {(Promise<void | InfrastructureComponentSummary>)}
+   * @memberof DataService
+   */
   public async getComponentDetailData(id: string, date?: DateTime): Promise<void | InfrastructureComponentSummary> {
 
     let suffix = '?id=' + encodeURIComponent(id);
@@ -90,23 +128,17 @@ export class DataService {
     return axios
       .get(url)
       .then((response) => response.data)
-      .then((data) => processData(data))
-      //.catch((error) => console.log(error)) ;
-
-    // TODO: move this method to InfrastructureComponentSummary() 
-    function processData(data): InfrastructureComponentSummary {
-      let result = <InfrastructureComponentSummary>data;
-      if (result.component.category === 'Subscription') {
-        result.component.category = 'Azure Subscription';
-      }
-      result.scoreHistory = result.scoreHistory.reverse(); //.slice(0, 14);
-      result.current = new CountersSummary(data.current);
-      console.log(`[] result`, result);
-      return result;
-    }
-
+      .then((data) => InfrastructureComponentSummary.fromData(data))
   }
 
+  /**
+   * Return image scan result data.
+   *
+   * @param {string} imageTag
+   * @param {string} date
+   * @returns {(Promise<void | ImageScanDetailModel>)}
+   * @memberof DataService
+   */
   public async getImageScanResultData(imageTag: string, date: string): Promise<void | ImageScanDetailModel> {
     const suffix = this.fixedEncodeURIComponent(imageTag) + '/details/?date=' + date + '&api-version=' + this.apiVersion;
     const url = this.baseUrl + "/audits/container-image/" + suffix;
@@ -115,59 +147,17 @@ export class DataService {
     return axios
       .get(url)
       .then((response) => response.data)
-      .then((data) => processData(data))
-      //.catch((error) => console.log(error));
-
-    function processData(data: any): ImageScanDetailModel {
-      console.log(`[] processing`, data);
-      let result = new ImageScanDetailModel();
-      result.date = data.date.split('T')[0];
-      result.description = data.description
-      result.scanResult = data.scanResult
-      result.image = data.image
-
-      try {
-
-        for (let i = 0; i < data.targets.length; i++) {
-          let target = new TargetGroup(data.targets[i].target);
-
-          for (let j = 0; j < data.targets[i].vulnerabilities.length; j++) {
-            let vulnerability = data.targets[i].vulnerabilities[j];
-
-            // split the references if it is not splitted correctly
-            if (vulnerability.references.length === 1) {
-              let newReferencesArray = vulnerability.references.slice()[0].split('\n');
-              vulnerability.references = [];
-              for (let r = 0; r < newReferencesArray.length - 1; r++) {
-                vulnerability.references.push(newReferencesArray[r])
-              }
-            }
-
-            let index = target.vulnerabilities.findIndex((v) => v.Severity === vulnerability.severity);
-            if (index < 0) {
-              let vulgroup = new VulnerabilityGroup(vulnerability.severity);
-              vulgroup.Count = 1;
-              vulgroup.Order = ScoreService.getOrderBySeverity(vulnerability.severity);
-              vulgroup.CVEs.push(vulnerability);
-              target.vulnerabilities.push(vulgroup);
-            } else {
-              target.vulnerabilities[index].CVEs.push(vulnerability);
-              target.vulnerabilities[index].Count += 1;
-            }
-          }
-          target.vulnerabilities.sort((a, b) => a.Order > b.Order ? -1 : a.Order < b.Order ? 1 : 0);
-          result.targets.push(target);
-        }
-
-      } catch (e) {
-        console.log(`error parsing image scan detail data ${e}`)
-      }
-      console.log(`[] result`, result);
-
-      return result;
-    }
+      .then((data) => ImageScanDetailModel.fromData(data))
   }
 
+  /**
+   * Return general overview diff data.
+   *
+   * @param {string} date1
+   * @param {string} date2
+   * @returns {(Promise<void | InfrastructureOverviewDiff>)}
+   * @memberof DataService
+   */
   public async getGeneralOverviewDiffData(date1: string, date2: string): Promise<void | InfrastructureOverviewDiff> {
     let suffix = '?date1=' + date1 + '&date2=' + date2  + '&api-version=' + this.apiVersion;
     let url = this.baseUrl + "/audits/overview/diff" + suffix;
@@ -176,60 +166,35 @@ export class DataService {
     return axios
       .get(url)
       .then((response) => response.data)
-      .then((data) => processData(data))
-      //.catch((error) => console.log(error));
-
-    function processData(data:any) : InfrastructureOverviewDiff {
-      console.log(`[]input`, JSON.parse(JSON.stringify(data)));
-      let result = new InfrastructureOverviewDiff();
-      result.summary1 = InfrastructureOverview.GenerateFromDiff(data.overall1, data.components1);
-      result.summary2 = InfrastructureOverview.GenerateFromDiff(data.overall2, data.components2);
-      result.compositeComponents = [];
-      
-      let cc:InfrastructureComponentSummary[] = []
-
-      for(let i=0;i<result.summary1.components.length;i++) {
-        let id = result.summary1.components[i].component.id;
-        let index = cc.findIndex(x=>x.component.id === id);
-        if(index === -1) {
-          cc.push(result.summary1.components[i]);
-        }
-      }
-      for(let i=0;i<result.summary2.components.length;i++) {
-        let id = result.summary2.components[i].component.id;
-        let index = cc.findIndex(x=>x.component.id === id);
-        if(index === -1) {
-          cc.push(result.summary2.components[i]);
-        }
-      }
-
-      result.compositeComponents = cc;   
-      console.log(`[]output`, JSON.parse(JSON.stringify(result)));      
-      return result;
-    }
+      .then((data) => InfrastructureOverviewDiff.fromData(data))
   }
 
+  /**
+   * Return component history data.
+   *
+   * @param {string} id
+   * @returns {(Promise<void | InfrastructureComponentSummary[]>)}
+   * @memberof DataService
+   */
   public async getComponentHistoryData(id: string): Promise<void | InfrastructureComponentSummary[]> {
     let suffix = '?id=' + encodeURIComponent(id)  + '&api-version=' + this.apiVersion;
     let url = this.baseUrl + "/audits/component/history" + suffix;
     console.log(`[] calling ${url}`);
     return axios
       .get(url)
-      .then((response) => { 
-        console.log(response.data);
-        return response.data;
-      })
-      .then((data) => processData(data))
-      //.catch((error) => console.log(error))
-      //.finally(() => console.log("component history request finished."));
-
-    function processData(data): InfrastructureComponentSummary[] {
-      let result = data.reverse();
-      console.log(`[] result`, result);
-      return result;
-    }
+      .then((response) => response.data.reverse())
+      .then((data) => <InfrastructureComponentSummary[]>data)
   }
 
+  /**
+   * Return component diff data.
+   *
+   * @param {string} id
+   * @param {string} date1
+   * @param {string} date2
+   * @returns {(Promise<void | InfrastructureComponentDiff>)}
+   * @memberof DataService
+   */
   public async getComponentDiffData(id: string, date1: string, date2: string): Promise<void | InfrastructureComponentDiff> {
     let suffix = '?id=' + encodeURIComponent(id) + '&date1=' + date1 + '&date2=' + date2 + '&api-version=' + this.apiVersion;
     let url = this.baseUrl + "/audits/component/diff" + suffix;
@@ -238,22 +203,27 @@ export class DataService {
     return axios
       .get(url)
       .then((response) => response.data)
-      .then((data) => {
-        console.log(JSON.parse(JSON.stringify(data)));
-        return InfrastructureComponentDiff.CreateFromData(data)
-      })
-      //.catch((error) => console.log(error));
-
+      .then((data) => InfrastructureComponentDiff.fromData(data));
   }
 
+  /**
+   * Return website meta data.
+   *
+   * @returns {(Promise< void | MetaData[]>)}
+   * @memberof DataService
+   */
   public async getWebsiteMeta(): Promise< void | MetaData[]> {
     let url = this.baseUrl + "/knowledgebase/website-metadata?api-version=" + this.apiVersion;
     console.log(`[] calling ${url}`);
     return axios
       .get(url)
-      .then((response) => response.data)
-      .then((data) => data)
-      //.catch((error) => console.log(error));
+      .then((response) => <MetaData[]>response.data);
+  }
+
+  public fixedEncodeURIComponent(str: string) {
+    return encodeURIComponent(str).replace(/[!*]/g, function (c) {
+      return "%" + c.charCodeAt(0).toString(16);
+    });
   }
 
 }
